integration/requirements.txt
marionette_client==3.1.0
firefox-puppeteer==50.0.0
mozlog==3.3
